UserRegistrationController class renamed to SignUpController
