+++ src/main/java/com/webshoppe/services/vo/SignUpVO.java
@@ -1,0 +1,14 @@
+package com.webshoppe.services.vo;
+
+public class SignUpVO extends BaseVO{
+
+    private String userUUID;
+
+    public String getUserUUID() {
+        return userUUID;
+    }
+
+    public void setUserUUID(String userUUID) {
+        this.userUUID = userUUID;
+    }
+}
