+++ src/test/java/com/webshoppe/controllers/SignUpControllerTest.java
@@ -19,7 +19,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class UserRegistrationControllerTest {
+public class SignUpControllerTest {
 
     @Autowired
     private WebApplicationContext wac;
